refactor: _scalar moves to its own module for a second caller
